+++ domains/certificates/Certificates.sln
@@ -17,13 +17,13 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "API.UnitTests", "Query.API\
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MeasurementEvents", "Shared\MeasurementEvents\MeasurementEvents.csproj", "{7CE965F2-DF46-4372-A44F-BA14957DC9C9}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Contracts", "Shared\Contracts\Contracts.csproj", "{03A0CD09-F291-4D30-9D8A-A2DE18142312}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Contracts", "Shared\Contracts\Contracts.csproj", "{03A0CD09-F291-4D30-9D8A-A2DE18142312}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AggregateRepositories", "Shared\AggregateRepositories\AggregateRepositories.csproj", "{0020310F-7891-4C33-9490-A48F8BC09981}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AggregateRepositories", "Shared\AggregateRepositories\AggregateRepositories.csproj", "{0020310F-7891-4C33-9490-A48F8BC09981}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "RegistryConnector", "RegistryConnector", "{3F085CB6-318D-4E46-89DB-DBABA9277661}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "RegistryConnector", "RegistryConnector", "{9AB85139-82C5-45D0-97CD-CD9FCADB6380}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RegistryConnector", "RegistryConnector\RegistryConnector\RegistryConnector.csproj", "{7131CF14-F319-49A8-AEE8-0D8F455A8F31}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Worker", "RegistryConnector\Worker\Worker.csproj", "{005E4EDB-9F14-4CDC-A2A9-C5D261872E5C}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -59,10 +59,10 @@ Global
 		{0020310F-7891-4C33-9490-A48F8BC09981}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{0020310F-7891-4C33-9490-A48F8BC09981}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{0020310F-7891-4C33-9490-A48F8BC09981}.Release|Any CPU.Build.0 = Release|Any CPU
-		{7131CF14-F319-49A8-AEE8-0D8F455A8F31}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7131CF14-F319-49A8-AEE8-0D8F455A8F31}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7131CF14-F319-49A8-AEE8-0D8F455A8F31}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7131CF14-F319-49A8-AEE8-0D8F455A8F31}.Release|Any CPU.Build.0 = Release|Any CPU
+		{005E4EDB-9F14-4CDC-A2A9-C5D261872E5C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{005E4EDB-9F14-4CDC-A2A9-C5D261872E5C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{005E4EDB-9F14-4CDC-A2A9-C5D261872E5C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{005E4EDB-9F14-4CDC-A2A9-C5D261872E5C}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -75,7 +75,7 @@ Global
 		{7CE965F2-DF46-4372-A44F-BA14957DC9C9} = {50B52CB7-AA59-483B-8C89-46A681B3E01F}
 		{03A0CD09-F291-4D30-9D8A-A2DE18142312} = {50B52CB7-AA59-483B-8C89-46A681B3E01F}
 		{0020310F-7891-4C33-9490-A48F8BC09981} = {50B52CB7-AA59-483B-8C89-46A681B3E01F}
-		{7131CF14-F319-49A8-AEE8-0D8F455A8F31} = {3F085CB6-318D-4E46-89DB-DBABA9277661}
+		{005E4EDB-9F14-4CDC-A2A9-C5D261872E5C} = {9AB85139-82C5-45D0-97CD-CD9FCADB6380}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {8710F924-9EE6-4142-A232-456121E239B0}
