+++ domains/certificates/RegistryConnector/Worker/Properties/launchSettings.json
