+++ domains/certificates/RegistryConnector/Worker/Worker.csproj
@@ -1,4 +1,4 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.Web">
+<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
     <TargetFramework>net7.0</TargetFramework>
@@ -7,6 +7,8 @@
     <UserSecretsId>c55e4dee-e6a8-4b4a-aadb-61d75fb92b03</UserSecretsId>
     <DockerDefaultTargetOS>Linux</DockerDefaultTargetOS>
     <DockerfileContext>..\..</DockerfileContext>
+    <RootNamespace>RegistryConnector.Worker</RootNamespace>
+    <AssemblyName>RegistryConnector.Worker</AssemblyName>
   </PropertyGroup>
 
   <ItemGroup>
